day 16 part 2
test cases don't work for part 2. had to work it out by hand on the scratch paper

+++ src/test/java/adventOfCode2020/day16/Puzzle16Test.java
@@ -42,7 +42,7 @@ public class Puzzle16Test {
     private void assertSolve2(boolean isExampleData) {
         AbstractPuzzle puzzle = isExampleData ? puzzleExampleData : puzzleRealData;
         Object answerObj = puzzle.solve2();
-        int answer = (Integer) answerObj;
+        long answer = (Long) answerObj;
         long expected = puzzle.getExpectedAnswer2();
         assertTrue("Expected " + expected + " but got " + answer, answer == expected);
     }
